+++ src/debugger.rs
@@ -1,6 +1,6 @@
 use crate::context::run_with_plugin;
 use crate::define_data_plugin;
-use crate::extension_api::{
+use crate::external_api::{
     run_extension, EmptyArgs, GlobalPropertyExtension, GlobalPropertyExtensionArgs,
     GlobalPropertyExtensionRetval, NextCommandExtension, NextExtensionArgs, PopulationExtension,
     PopulationExtensionArgs,
