+++ src/lib.rs
@@ -55,5 +55,5 @@ pub use runner::{run_with_args, run_with_custom_args, BaseArgs};
 
 pub mod debugger;
 
-pub mod extension_api;
+pub mod external_api;
 pub mod web_api;
